Drop fixture from plugin conftest
