+++ saleor/plugins/tests/fixtures.py
@@ -106,17 +106,3 @@ def vatlayer(db, tax_rates, taxes, setup_vatlayer):
     }
     VAT.objects.create(country_code="DE", data=tax_rates_2)
     return taxes
-
-
-@pytest.fixture
-def user_email_plugin(settings):
-    def fun():
-        settings.PLUGINS += ["saleor.plugins.user_email.plugin.UserEmailPlugin"]
-        manager = get_plugins_manager()
-        manager.save_plugin_configuration(
-            UserEmailPlugin.PLUGIN_ID, {"active": True, "configuration": []}
-        )
-        manager = get_plugins_manager()
-        return manager.plugins[0]
-
-    return fun
